[Model] Prevent duplicate categories
On a Mac, Apple Music creates a nested ~/Music/Music dir. If you keep
your mpd music in that directory, you end up with two Music categories,
which was triggering a JS bug that I haven't properly fixed. Because of
the way we find categories, the second one isn't even usable (that's its
own problem, I guess) so the best thing to do is just skip adding a
category if another one exists with the same name.