[Rust] Basic actix & rodio player
A working PoC of an Actix server with a Rodio player running in a background
thread,  communicating over a channel.

This design came from trying a few things, but with Rodio, we need to
keep a reference to the output stream, and that is safe to move across
threads, and does not implement `Send`. Therefore, we have this setup
where a single thread owns the Rodio data and we pass commands to it
over a channel.

There are plenty of things to improve even before we build out the
filesystem and control systems, the first of which being a typed command
to send over the channel, but this initial commit is a good checkpoint.