Better type for Sender
Instead of sending a string, send something that can be better typed.
Problem: We could send an array of String arguments with a command, but
they aren't guaranteed to match the command (e.g. play requires one filename)
and if we have to pass a non-string type we'll be parsing and hoping for the
best.

Solution: create a Command trait that accepts a `Player` to call whatever method
we like on the player from within the Player thread. The channel needs to send
it as a `Box<dyn Command>` and the trait needs to implement `Send`.

- Implement PlayCommand, StopCommand, VolumeCommand
- Implement Volume endpoint to prove it out

Also, drop the mutex in app state since we are cloning the sender, each
thread gets it's own sender.