[Build] Parallelize rust test and build
Split fmt/clippy/test (debug builds) from build (release) to improve
caching. It looks like with this setup the caching of .target/ leads to
a significant speedup, plus these two flows can run in parallel.

In testing we see a change from 8 mins to 1 min, but this may be
distorted by having a 'hot' actions host, we may not see this for
infrequent builds.