[CI] Drop rust-toolchain and rust-cache actions
The actions-rs/toolchain action is poorly maintained, and it looks like it
might not be needed.

The rust-cache action was somehow calculating a different key to save
the cache than the one it used to restore, so there was never a cache
hit. Not sure why. Switching to the plain `actions/cache` seems to fix
the problem (build time from 6min to under 1min).

Also, set the working-directory for all steps as a default.